+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: gsilva <gsilva@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/13 10:48:43 by gsilva            #+#    #+#             */
-/*   Updated: 2022/10/07 11:25:30 by gsilva           ###   ########.fr       */
+/*   Updated: 2022/10/13 09:35:48 by gsilva           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = ft_strlen((char *)s);
 	while (s[i] != c && i > 0)
